Change inEnv to custom operator
