+++ src/Inference/TypingEnvironment.hs
@@ -4,11 +4,14 @@ module Inference.TypingEnvironment where
 
 import Inference.Type
 
-import Prelude hiding (lookup)
-import AbsSyntax
+import Syntax.Base
 
-import Data.Monoid
-import Data.Semigroup
+import Prelude hiding (lookup)
+
+import Control.Monad.Except
+import Control.Monad.State
+import Control.Monad.Reader
+import Control.Monad.Identity
 import Data.Foldable hiding (toList)
 import qualified Data.Map as Map
 
@@ -19,6 +22,23 @@ data TypeErrorPayload = EmptyPayload | TypeErrorPayload String deriving (Show)
 type Constraint = (Type, Type)
 data AConstraint = AConstraint TypeErrorPayload Constraint
 type Unifier = (Subst, [AConstraint])
+
+
+type Infer = StateT (InferState) (ReaderT (Env) (ExceptT TypeError IO))
+data InferState = InferState { count :: Int, inferTrace :: [String], lastInferExpr :: String }
+
+data TypeError
+  = UnificationFail TypeErrorPayload Type Type
+  | InfiniteType TypeErrorPayload TVar Type
+  | UnboundVariable TypeErrorPayload Ident
+  | Ambigious TypeErrorPayload [Constraint]
+  | UnificationMismatch TypeErrorPayload [Type] [Type]
+  | Debug TypeErrorPayload String
+  deriving (Show)
+
+
+initInfer :: InferState
+initInfer = InferState { count = 0, inferTrace = [], lastInferExpr = "" }
 
 constraintToStr :: Constraint -> String
 constraintToStr (a,b) = (typeToStr [] a) ++ " ~ " ++ (typeToStr [] b)
@@ -34,6 +54,11 @@ empty = TypeEnv Map.empty
 
 (-->) :: Env -> Ident -> Env
 (-->) (TypeEnv env) var = TypeEnv (Map.delete var env)
+
+(==>) :: (Ident, Scheme) -> Infer a -> Infer a
+(==>) (x, sc) m = do
+  let scope e = (e --> x) ++> (x, sc)
+  local scope m
 
 extends :: Env -> [(Ident, Scheme)] -> Env
 extends env xs = env { types = Map.union (Map.fromList xs) (types env) }
